Consolidate inject/extract context logic into extensions
