+++ src/DurableOrchestrator/GlobalUsings.cs
@@ -2,6 +2,7 @@ global using System;
 global using System.Text.Json;
 global using System.Text.Json.Serialization;
 global using System.Threading.Tasks;
+global using DurableOrchestrator.Observability;
 global using Microsoft.Azure.Functions.Worker;
 global using Microsoft.Azure.Functions.Worker.Http;
 global using Microsoft.DurableTask;
